+++ custom_components/openspeedtest_cli/button.py
@@ -10,6 +10,8 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from . import OpenSpeedTestConfigEntry, build_device_info
 from .coordinator import OpenSpeedTestCoordinator
 
+PARALLEL_UPDATES = 0
+
 
 async def async_setup_entry(
     hass: HomeAssistant,